requirements.txt
streamlit
pdfplumber
python-docx
astrapy
requests
python-dotenv